+++ qatrack/parts/urls.py
@@ -1,5 +1,4 @@
-
-from django.conf.urls import include, url
+from django.conf.urls import url
 
 from qatrack.parts import views
 
@@ -9,7 +8,8 @@ urlpatterns = [
     url(r'^room_location_searcher$', views.room_location_searcher, name='room_location_searcher'),
     url(r'^list(?:/(?P<f>\S+))?/$', views.PartsList.as_view(), name='parts_list'),
     url(r'^new/$', views.PartUpdateCreate.as_view(), name='part_new'),
-    url(r'^edit/(?P<pk>\d+)?$', views.PartUpdateCreate.as_view(), name='part_edit'),
-    url(r'^details/(?P<pk>\d+)?$', views.PartDetails.as_view(), name='part_details'),
+    url(r'^edit/(?P<pk>\d+)?/$', views.PartUpdateCreate.as_view(), name='part_edit'),
+    url(r'^details/(?P<pk>\d+)?/$', views.PartDetails.as_view(), name='part_details'),
     url(r'^suppliers/$', views.SuppliersList.as_view(), name='suppliers_list'),
+    url(r'^supplier/details/(?P<pk>\d+)/$', views.SupplierDetails.as_view(), name='supplier_details'),
 ]